requirements.txt
starkbank==2.28.1
faker==37.1.0
python-dotenv==1.0.1